update to latest hyper, futures libs
These changes required a bit of refactoring and debugging, as the newly
async method resolve_conflicts did not play well with the rest of the
implementation. The issue stemmed from the fact that RwLock did not work
with futures, and the guard did not implement "Send", causing issues
with we were trying to lock the blockchain in a handler that had async
function calls within it.

This is resolved by switching the RwLock to a futures::lock::Mutex,
based on the recommendations in this SO post:
https://stackoverflow.com/a/67277503/2528719

The debug macros were added as it was rather difficult to figure out
what was going on without them, inspired by
https://github.com/tokio-rs/axum/discussions/641#discussioncomment-1843825